Removed dashboard code from main
The URL shortener dashboard now is only a subcomponent of the whole app,
meaning that it shouldn't really be part of the main file.
